export price interfaces in api.model file
