+++ projects/front-end-library/src/lib/components/card/angular/card.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewEncapsulation, OnInit, Input, Output, EventEmitter} from "@angular/core";
-import { IPrice } from '../../price/angular/price.component';
+import { IPrice } from '../../price/angular/api.model';
 
 
 interface IImage {
