+++ projects/front-end-library/src/lib/components/hero/angular/hero.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewEncapsulation, OnInit, Input } from '@angular/core';
-import { IPrice } from '../../price/angular/price.component';
+import { IPrice } from '../../price/angular/api.model';
 
 /**
  * Description of the component Hero
